refactor: 카테고리, 태그카테고리 캐싱을 위해 CategoryCacheService로 분리 및 caffeine 의존성 주입
